[Add]RotObjectSpin.csを追加しました。 RotatableObjectのまわす大についてのスクリプトです。
[Add]RotObjectRotate.csを追加しました
RotatableObjectのまわす小についてのスクリプトです。

[Add]Pf_BaceObjectを追加しました
基本オブジェクト用のプレハブを追加しました。

[Update]RotatableObject.csを一部変更しました
partial修飾子を追加してスクリプトをわけています。
