+++ Assets/Prefabs/Pf_BaceObject.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 210226ce7593c1b4f8aedceea7a6e29d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
