+++ Assets/Scripts/StageObjectScript/RotObjectRotate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 83349b2fca910374486b459aa120f7bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
