+++ Assets/Scripts/StageObjectScript/RotObjectSpin.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93ae8a3a596d8d54392b8f5563eda56a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
